feat: Add JSON/CSV storage option for scraped data
This commit adds the functionality to store scraped data in JSON or CSV format. Users can now choose to save the scraped data in either a `.json` or `.csv` file by using the appropriate command. This feature enhances the flexibility of the scraper and allows users to easily manage and analyze the scraped data in their preferred format.