feat: Add SQLAlchemyPipeline for storing scraped book data
This commit adds a new SQLAlchemyPipeline class to the pipelines.py file. The SQLAlchemyPipeline is responsible for storing the scraped book data in a MySQL database using SQLAlchemy. It provides methods for opening and closing the database session, as well as processing each scraped item and adding it to the database. This addition enhances the scraper's functionality by improving the storage and retrieval of book data.